+++ src/app-container.ts
@@ -10,6 +10,8 @@ import { CommandHandler } from '@app/processing/command-handler';
 import { RegisterUserCommandHandler } from '@app/user-access/register-user/register-user.command-handler';
 import { CommandBus } from '@app/processing/command-bus';
 import { MailerService } from '@infrastructure/mailer/mailer.service';
+import { DomainSubscriber } from '@core/shared/domain-subscriber';
+import { UserRegisteredSubscriber } from '@app/user-access/register-user/user-registered.subscriber';
 
 const registerAsArray = <T>(resolvers: Awilix.Resolver<T>[]): Awilix.Resolver<T[]> => ({
   resolve: (container: Awilix.AwilixContainer) => resolvers.map((r) => container.build(r)),
@@ -54,6 +56,12 @@ export const createAppContainer = async (): Promise<Awilix.AwilixContainer> => {
 
   container.register({
     userRegistrationRepository: Awilix.asClass(UserRegistrationRepositoryImpl).singleton(),
+  });
+
+  container.register({
+    subscribers: registerAsArray<DomainSubscriber<any>>([
+      Awilix.asClass(UserRegisteredSubscriber).singleton(),
+    ]),
   });
 
   const app = container.resolve<Server>('server').getApp();