+++ src/index.ts
@@ -4,6 +4,7 @@
 import { createAppContainer } from './app-container';
 import { Logger } from '@infrastructure/logger/logger';
 import { Application } from 'express';
+import { DomainSubscriber } from '@core/shared/domain-subscriber';
 
 require('dotenv').config();
 
@@ -16,6 +17,7 @@ if (process.env.NODE_ENV === 'production') {
 
   const logger = container.resolve<Logger>('logger');
   const app = container.resolve<Application>('app');
+  const subscribers = container.resolve<DomainSubscriber<any>[]>('subscribers');
 
   process.on('uncaughtException', (error) => {
     logger.error(`Uncaught Exception: ${error.toString()}`, error);
@@ -30,6 +32,8 @@ if (process.env.NODE_ENV === 'production') {
   });
 
   const port = process.env.PORT;
+
+  subscribers.forEach((subscriber) => subscriber.setupSubscriptions());
 
   app.listen({ port }, () => {
     logger.info(`Server is listening on ${process.env.PROTOCOL}://${process.env.HOST}:${port}`);